requirements.txt
Pillow
requests
litemapy
